Remove redux from MFA screen
